Move report genration into separate module
