+++ src/yatm_v2/src/helpers/mod.rs
@@ -19,3 +19,5 @@ pub use test_case_to_markdown::{
     permutation_to_labels, project_version_to_label, sanitize_label, test_case_to_markdown,
 };
 pub use test_cases_builder::{get_test_cases_builders_from_file, validate_test_cases_builder_file};
+pub mod metrics;
+
